requirements.txt
bs4
requests
openai
python-dotenv
